move views and serializers into their own folders
