+++ resources_portal/views/user.py
@@ -1,9 +1,9 @@
 from rest_framework import mixins, viewsets
 from rest_framework.permissions import AllowAny
 
-from .models import User
-from .permissions import IsUserOrReadOnly
-from .serializers import CreateUserSerializer, UserSerializer
+from ..models import User
+from ..permissions import IsUserOrReadOnly
+from ..serializers import CreateUserSerializer, UserSerializer
 
 
 class UserViewSet(mixins.RetrieveModelMixin, mixins.UpdateModelMixin, viewsets.GenericViewSet):
